Reorganize and simplify web/server files
