+++ web/server/src/index.js
@@ -1,16 +1,12 @@
 import express from 'express'
 import bodyParser from 'body-parser'
-import testcaseRoutes from './routes/testcase'
-import testRoutes from './routes/test'
-import utilityRoutes from './routes/utility'
+import boptestRoutes from './routes/boptestRoutes'
 
 var app = express()
 
 app.use(bodyParser.urlencoded())
 app.use(bodyParser.json())
-app.use('/', testcaseRoutes)
-app.use('/', testRoutes)
-app.use('/', utilityRoutes)
+app.use('/', boptestRoutes)
 
 let server = app.listen(80, () => {
   var host = server.address().address
